Fixed #1.
The MiniMap canvas belonged to the Default layer. So Added Default layer to the capture target.

In addition, I made one misunderstanding.

I was currently forced to enable FPS mode. The MiniMap disappears from the screen when in FPS mode.
Therefore, the map is normally invisible.

If you turn off FPS mode, you can see the map. FPS mode can be canceled with the space key. This is standard operation.